refactor(worker): extract utility functions to dedicated module
Address PR review feedback by exporting utility functions rather than
reimplementing them in tests to prevent code divergence.

- Create utils.ts with all testable utility functions
- Update index.ts to import from utils.ts
- Update tests to import actual functions instead of reimplementing
- All 150 tests continue to pass
